Make WORKING Many-to-Many relationship for Waiter and Rat, create, update and delete operations included. Also remove one of DTOs, because it was overkill.

+++ src/modules/waiter/services/waiter/waiter.service.ts
@@ -5,13 +5,13 @@ import { CreateWaiterDto } from "../../dto/create-waiter.dto";
 import { UpdateWaiterDto } from "../../dto/update-watier.dto";
 import { Waiter } from "../../entities/waiter.entity";
 import { Rat } from "src/modules/rat/entities/rat.entity";
-import { CreateWaiterWithRatsDto } from "../../dto/create-waiter-with-rats.dto";
 
 @Injectable()
 export class WaiterService {
     constructor(
         @InjectRepository(Waiter)
         private readonly waiterRepository: Repository<Waiter>,
+
         @InjectRepository(Rat)
         private readonly ratRepository: Repository<Rat>,
     ) {}
@@ -20,10 +20,27 @@ export class WaiterService {
         try {
             const waiter = this.waiterRepository.create(createWaiterDto);
 
-            return await this.waiterRepository.save(waiter);
-        } catch(err) {
-            Logger.error(`Error in create(): ${err.message}`);
-            throw err;
+            const ratIDs = createWaiterDto.ratIDs;
+
+            if (ratIDs) {
+                const rats = await this.ratRepository.findBy({id: In(ratIDs)});
+            
+                if (!Array.isArray(rats) || rats.length === 0) {
+                    Logger.error(`Error in create(): rat with IDs ${ratIDs} weren't found`);
+                    throw new EntityNotFoundError(Rat, { id: ratIDs });
+                }
+    
+                if (!waiter.rats) {
+                    waiter.rats = [];
+                }
+
+                waiter.rats.push(...rats);
+            }
+
+            return this.waiterRepository.save(waiter);
+        } catch (error) {
+            Logger.error(`Error in create(): ${error}`);
+            throw error;
         }
     }
 
@@ -32,6 +49,14 @@ export class WaiterService {
             return await this.waiterRepository.find();
         } catch(err) {
             Logger.error(`Error in findAll(): ${err.message}`);
+            throw err;
+        }
+    }
+
+    async findAllWaitersWithRats() {
+        try {
+            return await this.waiterRepository.find({relations: ['rats']});
+        } catch (err) {
             throw err;
         }
     }
@@ -47,13 +72,8 @@ export class WaiterService {
 
     async update(id: number, updateWaiterDto: UpdateWaiterDto): Promise<void> {
         try {
-            await this.waiterRepository.findOneByOrFail({ id })
-        } catch(err) {
-            Logger.error(`Error in update(): ${err.message}`);
-            throw err;
-        }
+            await this.waiterRepository.findOneByOrFail({ id });
 
-        try {
             await this.waiterRepository.update(id, updateWaiterDto);
         } catch(err) {
             Logger.error(`Error in update(): ${err.message}`);
@@ -61,17 +81,55 @@ export class WaiterService {
         }
     }
 
-    async remove(id: number): Promise<Waiter> {
-        let waiter: Waiter;
-
+    async addRatsToWaiter(id: number, ratIDs: number[]): Promise<Waiter> {
         try {
-            waiter = await this.waiterRepository.findOneByOrFail({ id })
+            const waiter = await this.waiterRepository.findOne({
+                relations: {
+                    rats: true,
+                },
+                where: { id: id }
+            })
+
+            const rats = await this.ratRepository.findBy({id: In(ratIDs)});
+
+            if (!Array.isArray(rats) || rats.length === 0) {
+                Logger.error(`Error in addRatsToWaiter(): rat with IDs ${ratIDs} weren't found`);
+                throw new EntityNotFoundError(Rat, { id: ratIDs });
+            }
+
+            if (waiter.rats === undefined) {
+                waiter.rats = [];
+            }
+
+            for (let i = 0; i < rats.length; i++) {
+                if (waiter.rats.indexOf(rats[i]) === -1) {
+                    waiter.rats.push(rats[i]);
+                }
+            }
+
+            return await this.waiterRepository.save(waiter);
         } catch(err) {
-            Logger.error(`Error in remove(): ${err.message}`);
             throw err;
         }
+    }
 
+    async remove(id: number): Promise<Waiter> {
         try {
+            const waiter = await this.waiterRepository.findOne({
+                relations: {
+                    rats: true,
+                },
+                where: { id: id }
+            })
+
+            if (!waiter) {
+                throw new EntityNotFoundError(Waiter, { id: id }); 
+            }
+
+            waiter.rats = [];
+
+            await this.waiterRepository.save(waiter);
+
             return await this.waiterRepository.remove(waiter);
         } catch(err) {
             Logger.error(`Error in remove(): ${err.message}`);
@@ -79,27 +137,38 @@ export class WaiterService {
         }
     }
 
-    async createWaiterWithRats(createWaiterWithRatsDto: CreateWaiterWithRatsDto) {
-        const ratIDs = createWaiterWithRatsDto.ratIDs;
-
-        const rats = await this.ratRepository.findBy({id: In(ratIDs)});
-        
-        if (!Array.isArray(rats) || rats.length === 0) {
-            Logger.error(`Error in createWaiterWithRats(): rat with IDs ${ratIDs} weren't found`);
-            Logger.error(new EntityNotFoundError(Rat, { id: ratIDs }));
-            throw new EntityNotFoundError(Rat, { id: ratIDs });
-        }
-
+    async removeRatsFromWaiter(id: number, ratIDs: number[]): Promise<Rat[]> {
         try {
+            const waiter = await this.waiterRepository.findOne({
+                relations: {
+                    rats: true,
+                },
+                where: { id: id }
+            })
 
-            const waiter = this.waiterRepository.create(createWaiterWithRatsDto);
+            if (!waiter) {
+                throw new EntityNotFoundError(Waiter, { id: id }); 
+            }
 
-            waiter.rats.push(...rats);
+            const ratsToBeDeleted = await this.ratRepository.findBy({id: In(ratIDs)});
 
-            return this.waiterRepository.save(waiter);
-        } catch (error) {
-            Logger.error(`Error in createWaiterWithRats(): ${error}`);
-            throw error;
+            if (!Array.isArray(ratsToBeDeleted) || ratsToBeDeleted.length === 0) {
+                Logger.error(`Error in removeRatsFromWaiter(): rat with IDs ${ratIDs} weren't found`);
+                throw new EntityNotFoundError(Rat, { id: ratIDs });
+            }
+
+            waiter.rats = waiter.rats.filter((rat) => {
+                if (!ratIDs.includes(rat.id)) {
+                    return rat
+                }
+            })
+
+            await this.waiterRepository.save(waiter);
+
+            return ratsToBeDeleted;
+        } catch(err) {
+            Logger.error(`Error in removeRatsFromWaiter(): ${err.message}`);
+            throw err;
         }
     }
 }
